// frontend/web/src/services/API/proserve-wb-provisioning-api/models/AvailableVersionDistribution.ts
export interface AvailableVersionDistribution {
    /**
     * Version ID
     * @type {string}
     * @memberof AvailableVersionDistribution
     */
    versionId: string;
    /**
     * Version name
     * @type {string}
     * @memberof AvailableVersionDistribution
     */
    versionName: string;
    /**
     * Version description
     * @type {string}
     * @memberof AvailableVersionDistribution
     */
    versionDescription?: string;
    /**
     * Recommended/not recommended flag
     * @type {boolean}
     * @memberof AvailableVersionDistribution
     */
    isRecommendedVersion: boolean;
    /**
     * 
     * @type {Array<ProductParameter>}
     * @memberof AvailableVersionDistribution
     */
    parameters?: Array<ProductParameter>;
    /**
     * Metadata of the product version.
     * @type {{ [key: string]: AvailableVersionDistributionMetadataItem; }}
     * @memberof AvailableVersionDistribution
     */
    metadata?: { [key: string]: AvailableVersionDistributionMetadataItem; };
    /**
     * Details of the components versions of the product version.
     * @type {Array<ComponentVersionDetail>}
     * @memberof AvailableVersionDistribution
     */
    componentVersionDetails?: Array<ComponentVersionDetail>;
    /**
     * Operating System version of the product version.
     * @type {string}
     * @memberof AvailableVersionDistribution
     */
    osVersion?: string;
}