// frontend/web/src/services/API/proserve-wb-provisioning-api/models/GetAvailableProductVersionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AvailableVersionDistribution } from './AvailableVersionDistribution';
import {
    AvailableVersionDistributionFromJSON,
    AvailableVersionDistributionFromJSONTyped,
    AvailableVersionDistributionToJSON,
} from './AvailableVersionDistribution';

/**
 * 
 * @export
 * @interface GetAvailableProductVersionsResponse
 */
export interface GetAvailableProductVersionsResponse {
    /**
     * 
     * @type {Array<AvailableVersionDistribution>}
     * @memberof GetAvailableProductVersionsResponse
     */
    availableProductVersions?: Array<AvailableVersionDistribution>;
}

/**
 * Check if a given object implements the GetAvailableProductVersionsResponse interface.
 */
export function instanceOfGetAvailableProductVersionsResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetAvailableProductVersionsResponseFromJSON(json: any): GetAvailableProductVersionsResponse {
    return GetAvailableProductVersionsResponseFromJSONTyped(json, false);
}

export function GetAvailableProductVersionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetAvailableProductVersionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'availableProductVersions': !exists(json, 'availableProductVersions') ? undefined : ((json['availableProductVersions'] as Array<any>).map(AvailableVersionDistributionFromJSON)),
    };
}

export function GetAvailableProductVersionsResponseToJSON(value?: GetAvailableProductVersionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'availableProductVersions': value.availableProductVersions === undefined ? undefined : ((value.availableProductVersions as Array<any>).map(AvailableVersionDistributionToJSON)),
    };
}