// frontend/web/src/services/API/proserve-wb-provisioning-api/models/ProvisioningParameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProvisioningParameter
 */
export interface ProvisioningParameter {
    /**
     * The parameter key.
     * @type {string}
     * @memberof ProvisioningParameter
     */
    key: string;
    /**
     * The parameter value.
     * @type {string}
     * @memberof ProvisioningParameter
     */
    value?: string;
}

/**
 * Check if a given object implements the ProvisioningParameter interface.
 */
export function instanceOfProvisioningParameter(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "key" in value;

    return isInstance;
}

export function ProvisioningParameterFromJSON(json: any): ProvisioningParameter {
    return ProvisioningParameterFromJSONTyped(json, false);
}

export function ProvisioningParameterFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProvisioningParameter {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'key': json['key'],
        'value': !exists(json, 'value') ? undefined : json['value'],
    };
}

export function ProvisioningParameterToJSON(value?: ProvisioningParameter | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'key': value.key,
        'value': value.value,
    };
}